move image logic to a separate file, api consts
